14869 Don’t set the loading state when refreshing
When we pull to refresh, the built in controller shows a loading indicator.

If we set our state to `loading`, we also showed a ghost cell at the bottom of the list.

On a pull to refresh, it’s unlikely that the list is going to get longer – more likely, the full first page of results will be shown, so it’ll be the same length or shorter.

This commit prevents us from showing the ghost cell on refresh.