14869 Declare refreshable modifiers on list views
Previously we had to put our refreshable modifiers higher in the heirarchy than the views they were refreshing, because we changed the state to `loading` when we pulled to refresh. This change of state cancelled the pull to refresh network task.

The state change for refresh was removed in 01cfd95fbd01e45e4a63982390787ab00bd15e84, and now we can group the refreshable control more closely with the list view it relates to.