15180 Clear order when exiting the POS
Before exiting point of sale, we warn the merchant about losing their in-progress order.

We need to actually clear it down as any accidental retention can cause issues especially when reconnecting card readers.

This is due to a memory leak which still needs to be fixed. Without this mitigation, multiple payments are attempted when connecting a reader, if you’ve got any leaked POS instances in memory with unpaid orders saved in them.