15180 Fix leak when leaving POS on iOS 17
In iOS 17, the `posModel` previously wasn’t reliably deallocated when leaving POS. This could lead to issues with multiple card payment services awaiting reader reconnection, and confusing payment behaviour when a reader is reconnected.

The issue turned out to be that the `NavigationStack` holding the `ItemListView` wasn’t reliably removed from the view heirarchy, and that held on to `posModel`, and by extension everything else.

It worked correctly on iOS 18, and also in iOS 17 on a simulator.

To fix it for iOS 17 on devices, we fall back to `NavigationView`, which doesn’t leak in this way.

We still want to use the newer `NavigationStack` going forwards, as `NavigationView` is deprecated.