Remove syncOrderItemSelectionStateOnDismiss when dismissing OrderForm
syncOrderItemSelectionStateOnDismiss is called in a single-view setup (on iPhone) when Product selection is dismissed.

However, it's not needed to synchronize selection state with order items when dismissing the whole Order view. It creates issues on split views, such as resetting the state when "Discard changes" alert appears. If the alert is dismissed, then the view is in reset state.