15223 Fix `Try again` for non-retry location error
The cancellation action from the non-retryable location error during payment never returns if the `paymentCancellable` isn’t valid.

That can happen when we call `collectPaymentMethod` with an intent which has already had a location permissions error.

By nilling the cancellable when we produce an error from the completion handler, second and subsequent retries of the location error will use the fallback approach and successfully cancel.