Ensure we consistently show the loading state
The current behaviour is to start a search immediately when a key is typed, so that we show the loading state. Further keypresses are debounced by 300ms, reducing the number of searches while the user is typing.

We have to keep track of whether a search finished, to  ensure that we correctly debounce new searches.