Keep search state when coming back from the cart
If you go to the cart from a search state, then come back, the search term now stays in place.

Note that this isn’t only when coming back from the cart, it’s also when starting a new order, which isn’t great.

However, before this change it would keep the choice to show the search, but not the search term, so you’d just see the wrong items controller or the empty state.

I’ll improve it further in future work.