Animate changes between item list states
Note that we no longer change the selected list type based on the search term, it’s only by tapping on the search button or the back button. That code clearly wasn’t really doing anything now that the search field isn’t rendered when we go back… or something.