Move WPMediaPicker from CocoaPods to Module by copying sources
WPMediaPicker, https://github.com/wordpress-mobile/MediaPicker-iOS, is a
legacy library that has been declared as no longer maintained in
November 2023,
https://github.com/wordpress-mobile/MediaPicker-iOS/commit/352bdc88d61d13571fe86dfc99cbb83913eb2a32.

Unfortunately, WooCommerce still depends on it in a few places, so that
dropping it is a more complex task than the scope of the CocoaPods
removal project that triggered this work, see
https://linear.app/a8c/issue/AINFRA-216.

I therefore decided to compromise and apply the same strategy as
WordPressAuthenticator and WordPressKit: fold the repository into the
application to do without the dependency dance.

Given it seems WooCommerce is the only remaining client for
WPMediaPicker, folding into the app is better for us than adding SwiftPM
support directly into the library.

The only downside is that to fold the library we also need to fold its
assets, adding a bunch of binary files Git is not optimized for
handling. Luckily, the assets are just a few KBs in total, so at least
this won't bloat the repo and affect the clone time.