Address FIXME in WooFoundation left during migration
In migrating from an iOS vs watchOS framework architecture to a
cross-platform package one, we had to change how the code deals with
accessing the `Bundle`. In particular, the WooFoundation `UIColor` code
is no longer in the same `Bundle` where the palette is.

I chose to expose the palette's `Bundle` via a property on
`ColorStudio`.  This makes the `Bundle` instance accessible in
WooFoundation so that WooFoundationCore can remain UIKit free and run on
watchOS.