Explicitly `import` Foundation, UIKit, and other required frameworks
When moving from frameworks to Swift packages, see for example
https://github.com/woocommerce/woocommerce-ios/pull/15622 and
https://github.com/woocommerce/woocommerce-ios/pull/15651, we lose the
framework headers which had statements like `#import
<Foundation/Foundation.h>`. This means that basic frameworks such as
Foundation and UIKit are no longer imported as part of consumers
frameworks such as Yosemite. To simplify the code review of the upcoming
PRs migrating those frameworks to Swift packages in `Modules/`, this PR
preemptively adds the required imports in the files that need them.

This was done by running `git reset trunk` on the workbench branch
`temp-branch-for-swiftpm-migration` and only committing the `import`
additions in this branch.