AINFRA-674 - Update Codegen for WooFoundation and Networking
Updates the Sourcery configuration to restore code generation after the
recent changes to WooFoundation and Networking.

**WooFoundation**

When updating WooFoundation to be a Swift package, see
https://github.com/woocommerce/woocommerce-ios/pull/15651, I neglected
updating the Codegen/Sourcery configuration.

In the migration from framework to package, we lost the Xcode project
file. Sourcery provides a `package` configuration option, but I was not
able to make it work.

The only option I could find was to list `sources`.

**Networking**

It seems that after migrating the Networking project to synchronized
folders, https://github.com/woocommerce/woocommerce-ios/pull/15653,
Sourcery could no longer access the sources.

The only fix I could find was to move from passing a `project` to
passing a straight `sources` array.

**`args.moduleName`**

Unfortunately using `sources` results in Sourcery being unable to add
the module name to the `globalName` property which we use in the
templates to identify each of the types handle. Including the module is
necessary to avoid name clashes. To work around this limitation, I
updated the template to check for an optional `moduleName` argument in
the configuration and interpolate it in the type name if present.