Remove maximum barcode scan time
We have not found examples of scanners which don’t send `\r\n` style terminators at the end of a scan. Processing a scan after a delay risks being more confusing and makes the code more complex, so we’ve decided not to do it any more.

This commit removes all the code related to the 1.5s timeout for scans