revert reloadsections from being async, move receipt eligibility to syncEverything
reloadSections() was async because it called await createPaymentSection(). This might create a race condition where by the time cells were configured, async operations had changed the underlying data, causing index out of bounds crashes when we attempt to render other rows that are not currently in view, like shippin lines.