Fix split view not appearing on iPad after the initial login
The fixTabBarTraitCollectionOnIpadForiOS18 hack was not working properly on first launch after login because it requires view.window to be available to access the window's trait collection.

During initial login flow:
  1. viewDidLoad() is called before tab controller is added to window hierarchy
  2. AppCoordinator.displayLoggedInUI() sets tab controller as window root
  3. view.window becomes available, but fixTabBarTraitCollectionOnIpadForiOS18
     was already called when window was nil

Fixed by adding fixTabBarTraitCollectionOnIpadForiOS18() call to viewDidAppear() which is guaranteed to run after the view controller is in the window hierarchy and view.window is available.