Treat coupons as expired when we reach expiry date
We have a test for `test_isExpired_when_current_date_then_returns_true`, however, we didn’t quite match that. We actually relied on a small amount of time elapsing between the call to create the POSCoupon with `Date.now` and the check for whether the coupon had expired.

In some test runs, there was no measurable time change, so the test would fail.

This fixes the implementation to match the test, and general understanding of an expiry time – when you reach the time, the coupon expires.