Fix failing date tests
Previously we used an unlikely timezone: GMT+08h08m22s. This was used to test that converted dates were the same as one another.

However, it seems that now, iOS ignores or rounds down the seconds when creating a timezone. That resulted in test failures where we compared the seconds to `0` – under iOS 26, the actual value was 22, equal to the number of seconds after the last whole minute in the timezone we used.

I’ve made it an exact number of minutes to fix the test.