Show error for initial catalog sync failure
If we’ve never synced the catalog successfully, and the sync fails, we show a full screen error allowing the user to retry the sync.

Errors on subsequent syncs aren’t shown, we continue to allow the user to use POS as we already have their data. If it gets too old, we’ll show a stale catalog warning.