Use upsert for incremental updates
This commit fixes an issue with incremental sync

Previously, if variations changed, any unchanged variations for the same product would be deleted during an incremental sync.

This was because we also get the parent product back, and we used `insert(onConflict: .replace)` – this triggered a cascade delete of variations (and other relations.)

Then we inserted the changed variations, so any that were unchanged remained deleted.

To avoid these issues, we move to `save`, which is GRDB’s upsert method.

Additionally, with the fix deleted variations would not be deleted during an incremental sync. To improve that, we check the `variationIDs` array for which variations to leave in, after an incremental sync.