Hide cellular toggle on non-cellular devices
Many iPads don’t have a cellular connection. It’s a little confusing to show the cellular toggle when it doesn’t apply.

Unfortunately, there’s no good (undeprecated) API for checking this. I’ve used an approach of listing the network interfaces and looking for the cellular one, but it’s undocumented and might not always work.

OTOH, this isn’t a critical feature, so we can keep an eye on it.