use-mocked-card-present-payment for UI tests
Despite tests passing, the contents of the pos-payment screenshot was incorrect, as lands on the failure case for `checkOrderIsStillEligibleForPayment` when attempting to retrieve the order details, hence showing a “Error: unable to process payment. we could not fetch the latest order details” rather than the payment view.

Despite passing the CPP mock in the POSTabCoordinator, we were never setting it in the test as a launch argument.