Ignore time based sync eligibility inside POS
When we set a max catalog age of zero in a refresh, we’re forcing a sync – e.g. because of a pull to refresh, or a full sync from settings, or after a purchase.

When we do this, we should ignore temporal reasons for not syncing. Otherwise, it’s possible to get into a position where syncs don’t work any more even though they should.