Specify variation_fields when generating catalog
There’s now a separation between product fields and variation fields, so we can fetch what we need and have more specificity about what’s in the response.

With this shift, the file format has changed to return products/variations wrapped with their type, so I’ve updated the JSON parsing to match that here.