Avoid using 0 as a coupon diffableDataSource ID
Using the same ID for multiple rows in a diffable datasource causes a crash. This is rare (affecting 15 users) but persistent (it’s happened 115 times over the last year.)

The ID is set by the API, so it shouldn’t really be 0 for anything that’s been saved, but internally PHP sets them up with 0 during creation, so perhaps some plugins cause these to be returned.

Code _should_ be unique, so is a reasonable fallback while still being stable within the coupon list.